// frontend/src/api/model/patientCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export interface PatientCreate {
  /**
     * @minLength 1
     * @maxLength 100
     */
  first_name: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  last_name: string;
  document_type: string;
  document_number: string;
  birth_date?: string | null;
}